# docs/swagger.yaml
definitions:
...
  dto.GetAccountRequest:
    description: 请求获取账户信息时所需参数
    properties:
      email:
        type: string
      nickname:
        type: string
      phone:
        type: string
      role_code:
        type: string
    required:
      - email
      - nickname
      - phone
      - role_code
    type: object
...
  dto.LoginRequest:
    description: 用户登录请求所需参数
    properties:
      email:
        type: string
      img_verification_code:
        type: string
      password:
        type: string
    required:
      - email
      - img_verification_code
      - password
    type: object
  dto.LoginResponse:
    description: 用户登录成功后返回的响应，包括用户ID和令牌
    properties:
      access_token:
        description: 访问令牌
        type: string
      refresh_token:
        description: 刷新令牌
        type: string
      user_id:
        description: 用户ID
        type: integer
    type: object
  dto.RegisterRequest:
    description: 用户注册所需参数
    properties:
      email:
        type: string
      email_verification_code:
        type: string
      img_verification_code:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
    required:
      - email
      - email_verification_code
      - img_verification_code
      - nickname
      - password
    type: object
  dto.ResetPwdRequest:
    description: 用户重置密码所需参数
    properties:
      again_new_password:
        maxLength: 20
        minLength: 6
        type: string
      email:
        type: string
      email_verification_code:
        type: string
      new_password:
        maxLength: 20
        minLength: 6
        type: string
    required:
      - again_new_password
      - email
      - email_verification_code
      - new_password
    type: object
...
  vo.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      requestId: {}
      timestamp: {}
    type: object
info:
  contact: {}
paths:
...
  /account/getAccount:
    post:
      consumes:
        - application/json
      description: 根据提供的邮箱获取对应用户的详细信息
      parameters:
        - description: 获取账户请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.GetAccountRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.GetAccountRequest"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
        "404":
          description: 用户不存在
          schema:
            $ref: "#/definitions/vo.Result"
      summary: 获取账户信息
      tags:
        - 账户
  /account/getUserProfile:
    post:
      description: 获取当前登录用户的个人信息
      produces:
        - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    additionalProperties: true
                    type: object
                type: object
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 从本地存储中获取用户信息
      tags:
        - 账户
  /account/loginAccount:
    post:
      consumes:
        - application/json
      description: 用户登录并获取访问令牌，支持图形验证码校验
      parameters:
        - description: 登录信息
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.LoginRequest"
        - description: 图形验证码
          in: query
          name: ImgVerificationCode
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: 登录成功，返回访问令牌
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.LoginResponse"
                type: object
        "400":
          description: 参数错误，验证码校验失败
          schema:
            $ref: "#/definitions/vo.Result"
        "401":
          description: 登录失败，凭证无效
          schema:
            $ref: "#/definitions/vo.Result"
      summary: 用户登录
      tags:
        - 账户
  /account/logoutAccount:
    post:
      description: 退出当前用户登录状态
      produces:
        - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    type: string
                type: object
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 用户登出
      tags:
        - 账户
  /account/registerAccount:
    post:
      consumes:
        - application/json
      description: 注册新用户账号，支持图形验证码和邮箱验证码校验
      parameters:
        - description: 注册信息
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.RegisterRequest"
        - description: 图形验证码
          in: query
          name: ImgVerificationCode
          required: true
          type: string
        - description: 邮箱验证码
          in: query
          name: EmailVerificationCode
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.RegisterRequest"
                type: object
        "400":
          description: 参数错误，验证码校验失败
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      summary: 用户注册
      tags:
        - 账户
...
  /authenticated/resetPassword:
    post:
      consumes:
        - application/json
      description: 重置用户账户密码，支持邮箱验证码校验
      parameters:
        - description: 重置密码信息
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.ResetPwdRequest"
        - description: 邮箱验证码
          in: query
          name: EmailVerificationCode
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: 密码重置成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    type: string
                type: object
        "400":
          description: 参数错误，验证码校验失败
          schema:
            $ref: "#/definitions/vo.Result"
        "401":
          description: 未授权，用户未登录
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 重置密码
      tags:
        - 账户
...
swagger: "2.0"